requirements.txt
streamlit
boto3
chromadb
PyMuPDF
google-generativeai
requests
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-groq
